Reformatting master_tt_dic
Master_t_dic is now made in such a way so that the only incorrect day
mapping is the last half of the dark cycle. In other words, 00:01 on
the 14th is actually mapped to the 13th, so that 23:59 on the 13th is
read in the same day as 00:01 on the 14th. This makes pulling one cycle
very easy.
IMPORTANT- There is no more “light only” file. There is no more need to
manipulate .csv files beforehand (they are cleaned and read
automatically.
Before- the first half of the dark cycle was on the right day, and the
rest was on the wrong day. You’d have to manually delete many rows of
the csv to get dict reader to work and copy and paste data to get the
day structure you wanted.